define separate module for image utilities
